+++ docs/releasenotes.rst
@@ -15,7 +15,7 @@ This is a new main release branch, TrackMe 1.2.x requires the deployment of the 
 TrackMe requires a summary index (defaults to trackme_summary) and a metric index (defaults to trackme_metrics):
 https://trackme.readthedocs.io/en/latest/configuration.html
 
-- Feature Issue #xxx - new REST API endpoints for Elastic Sources management
+- Feature Issue #xxx - new REST API endpoints for Elastic Sources
 
 Version 1.2.29
 ==============